refactor(tile): tile/mod.rs was four files in one
A `mod.rs` should declare modules. This one declared eleven and then held
1351 lines: the `Tile` struct and its ~40 methods, `TileKind`, the whole
`QuantInfo`/`DequantAt` vocabulary, and the procedural constructors.

Now `tile/base.rs` (the tile and the store it dispatches on),
`tile/quant.rs` (what a quantized store carries), and
`tile/procedural/tile.rs` (the constructors, beside the recipes they
build — `procedural_virtual` went with them, since only they call it).
`tile/mod.rs` is 29 lines of declarations.

`tile/view/mod.rs` drops its `pub mod`s for the private ones every other
module in the crate uses; that also un-shadows the new `tile::quant`.

Clippy caught the one thing the mechanical split got wrong: `DeliveryFamily`'s
doc comment had been orphaned in `physical/residence.rs`, away from its trait.

No behaviour change.
